Try to fix landscape games in "Cover game Zoom" and "portrait" games in Grid zoom not occupying whole image.
By setting width or height 5 times larger than needed and specifying the exact width or height for the other value.
Skraper will stop resizing if one the sides reaches the viewport borders. This will fix landscape games in the zoom cover and
Vertical games in the grid zoom (although it depends on how big the height was initially compared to width in these vertical games)